Create and format/style header
Added text to the header in the form of the logo and the three links at
the top of the page. Gave a dark blue background with white text for the
logo. Gave the header padding in order to have "white" space between the
content and the edge of the page.