Add placeholder for image next to initial signup
Added a placeholder in gray for a product image. Sitting in a flexbox
with the initial signup text in order to add visual flair to website and
the product itself. Used flex box so that the image can grow depending
on the user screen. Needs more modifying to make sure that the image
does not take up text space.